fix(daemon): classify Windows taskkill exit races
